requirements.txt
anthropic
openai
requests
beautifulsoup4
gspread
google-auth
python-docx
streamlit
python-dotenv
tavily-python